Fix rich cursor encoding.
If the depth of the server is less than 24, when we built the cursor, we were
expanding the smaller pixels to the lower bits of the new pixel instead of the
upper bits.  The result was that colors in the cursor were showing up as dark
grays.

This patch refactors the cursor blitting function to be more similar to the
rest of the blitting functions and ensures that the proper shift values are
used.