If you call one of the vnc_display_open* functions followed quickly by vnc_display_close, then the library segfaults.
The reason is that the vnc_display_open* functions register an idle callback to launch the coroutine, but if vnc_display_close is called before the idle loop has had a moment to run, this idle callback remains.  The coroutine later runs - in my case, after the plugin I'm writing has been unmapped from memory, with unhappy consequences.

The attached patch fixes this by remembering the callback ID and unregistering it if necessary.