Update cursor grab to have the following behavior:
For normal VNC servers, keyboard/mouse grab should only be activated when
ctrl-alt is clicked.  If the server is doing cursor offloading, we should use
that cursor.

For servers supporting PointerTypeChange, when in absolute mode, they should
behave as above.  When in relative mode, pointer grab should also be activated
on the first left-click in the window.  The cursor should *not* be hidden when
cursor grab is not activated.