Always emit signals from system coroutine.
When using threads to simulate coroutines, GTK gets very confused as signal
handlers are run in a different thread than the main loop when the signal
handlers aren't expecting it.

This patch changes all of the signals emitted from the VNC coroutine to use
a delay function to actually issue the signals.  This appears to fix the
GThread coroutine issue.