Add support for colourmap based pixel formats
The current code will force the server to use the native
client's pixel format if it ever sees a pixel format
with true_color_flag==0 (ie colourmap).  This is bad because
the client native pixel format is almost always 32bpp while
the main reason for using a colourmap is to get bandwidth
friendly 8bpp format. This makes GTK-VNC look bad in any
performance comparisons.

This patch implements full support for rendering from
colourmaps. It introduces a boxed data type to store the
colour map entries and wires this upto the framebuffer
interface.

For rendering, the BLT functions for the VncBaseFramebuffer
are enhanced to cope with colourmaps. Since colour maps
provide 16bit r,g,b triples, the BLT code is enhanced to
cope with 64-bit bpp pixels. This is technically required
even for non-colourmap entries since the VNC pixel format
protocol message allows for 16 bit r,g,b masks which
results in a 64-bit true colour pixel.

* src/vnccolormap.h, src/vnccolormap.c, src/Makefile.am: Add
  a VncColorMap boxed data type to hold the colour map entries.
* src/vncframebuffer.h, src/vncframebuffer.c: Add an API to
  the VncFrameBuffer interface to allow the colour map to be
  updated.
* src/vncconnection.h, src/vncconnection.c: Remove last case
  of vnc_ops and replace with new VncColorMap object
* src/vncdisplay.c: Don't override the default server pixel
  format, allow use of colour maps
* src/vncbaseframebuffer.c, src/vncbaseframebufferblt.h: Add
  support for rendering pixels using a colourmap