Use separate soname when building against GTK-3
The GTK-3 library is not ABI compatible with GTK-2. Thus when
GTK-VNC builds against GTK-3 it is not ABI compatible with a
build done against GTK-2. It is thus neccessasry to allow for
parallel install of GTK-2 and GTK-3 builds. This is done by
changing the library name, pkgconfig name and include file
locations

 GTK-2:

    libgtk-vnc-1.0.so
    gtk-vnc-1.0.pc
    /usr/include/gtk-vnc-1.0

 GTK-3:

    libgtk-vnc-2.0.so
    gtk-vnc-2.0.pc
    /usr/include/gtk-vnc-2.0

NB, when enabling GTK-3, the traditional python bindings will
be disabled. The new GObject Introspection support provide an
alternative python binding for users of GTK-3

* gtk-vnc-1.0.pc.in: Fix to only use GTK-2
* gtk-vnc-2.0.pc.in: New for GTK-3 support
* src/Makefile.am: Different sonames & include dirs when
  building against GTK3
* configure.ac: Disable python binding with GTK3
* README: Add note about GTK3
* Makefile.am: Install new pkgconfig file for GTK3