Disable extended key events if keymap is unknown
Currently if we cannot determine the GDK keymap, we default to
the XT keymap. This is not a good choice because is causes
definite breakage on several platforms. If the keymap cannot
be determined, it is safer to disable the extended key event
support completely & rely on traditional VNC keysyms.

Also rename the x2pc methods to be gdk2rfb, since that's what
the conversion is actually doing.